+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Dominoes path".
 - Tests for "Hidden word".
 - Tests for "Brackets, enhanced edition".
+- Tests for "These romans are crazy!".
 
 ## [1.14.0] - 2022-10-31
 ### Added